requirements.txt
Flask==3.1.0
xrpl-py==4.0.0
python-dotenv==1.0.1
Werkzeug==3.1.3